+++ frontends/martypc_pixels_desktop/src/egui/image.rs
@@ -24,43 +24,40 @@
 
     --------------------------------------------------------------------------
 
-    lib.rs
+    egui::image.rs
 
-    Main emulator core 
-
+    Routines for loading and manipulating GUI images.
 */
 
-pub mod devices;
+use egui::ColorImage;
+//use image;
 
-pub mod breakpoints;
-pub mod bus;
-pub mod bytebuf;
-pub mod bytequeue;
-pub mod config;
-pub mod cpu_common;
-pub mod cpu_808x;
-pub mod floppy_manager;
-pub mod file_util;
-pub mod interrupt;
-pub mod machine;
-pub mod machine_manager;
-pub mod memerror;
-pub mod rom_manager;
-pub mod sound;
-pub mod syntax_token;
-pub mod tracelogger;
-pub mod updatable;
-pub mod util;
+static LOGO_IMAGE: &[u8] = include_bytes!("../../../../assets/marty_logo_about.png");
 
-pub mod vhd;
-pub mod vhd_manager;
-pub mod videocard; // VideoCard trait
+// Support other images later if needed
+#[derive(Debug, PartialEq, Eq, Hash)]
+pub enum UiImage {
+    Logo,
+}
 
-pub mod cpu_validator; // CpuValidator trait
+pub fn get_ui_image(img_select: UiImage) -> ColorImage {
 
-#[cfg(feature = "arduino_validator")]
-#[macro_use]
-pub mod arduino8088_client;
-#[cfg(feature = "arduino_validator")]
-#[macro_use]
-pub mod arduino8088_validator;
+    match img_select {
+        UiImage::Logo => {
+            // This shouldn't fail since its all static
+            load_image_from_memory(LOGO_IMAGE).unwrap()
+        }   
+    }
+}
+
+fn load_image_from_memory(image_data: &[u8]) -> Result<ColorImage, image::ImageError> {
+    let image = image::load_from_memory(image_data)?;
+    let size = [image.width() as _, image.height() as _];
+    let image_buffer = image.to_rgba8();
+    let pixels = image_buffer.as_flat_samples();
+    Ok(ColorImage::from_rgba_unmultiplied(
+        size,
+        pixels.as_slice(),
+    ))
+}
+