+++ frontends/martypc_pixels_desktop/src/egui/performance_viewer.rs
@@ -68,6 +68,13 @@ impl PerformanceViewerControl {
             ui.label(egui::RichText::new(format!("{}", self.stats.backend)));
             ui.end_row();
 
+            ui.label("Build: ");
+            #[cfg(debug_assertions)]
+            ui.label(egui::RichText::new(format!("DEBUG")));
+            #[cfg(not(debug_assertions))]
+            ui.label(egui::RichText::new(format!("Release")));
+            ui.end_row();
+
             ui.label("Internal resolution: ");
             ui.label(egui::RichText::new(format!("{}, {}", 
                 self.video_data.render_w, 
@@ -105,7 +112,7 @@ impl PerformanceViewerControl {
             ui.label("Emulation time: ");
             ui.label(egui::RichText::new(format!("{}", ((self.stats.emulation_time.as_micros() as f64) / 1000.0))));
             ui.end_row();
-            ui.label("Render time: ");
+            ui.label("Framebuffer time: ");
             ui.label(egui::RichText::new(format!("{}", ((self.stats.render_time.as_micros() as f64) / 1000.0))));
             ui.end_row();
             ui.label("Gui Render time: ");