+++ CHANGELOG.md
@@ -1,4 +1,23 @@
 
+## [0.1.3](https://github.com/dbalsom/martypc/releases/tag/0.1.3) (2023-07-06)
+
+* Disabled window doubling if window would not fit on screen due to DPI scaling.
+* Decreased minimum window size
+* Disabled warpspeed config flag
+* Renamed 'autostart' config flag to 'auto_poweron' and fixed poweron logic.
+* Mapped Right Alt, Control and Shift to emulated Left Alt, Control and Shift.
+* Added UI warning when MartyPC is compiled in debug mode (unfortunately the default behavior of cargo build)
+* Replaced CGA composite simulation code with reenigne's color multiplexer algorithm, for more accurate colors and a 3x speed improvement.
+* Added contrast, phase and CGA type controls to composite adjustment window.
+* Update Pixels to 0.13
+* Update egui and egui-wgpu 0.22
+* Update winit to 0.29* 
+    * Winit 0.29 fixes reported keyboard issues with non-US keyboard layouts unable to type certain keys.
+    * Winit 0.29 fixes excessively high updates per second on some Linux builds
+    * Enabled Wayland support
+* Enable basic clipboard integration in debugger for copy/paste of breakpoints. This feature will be expanded.
+* Fork egui-winit 0.22 to manually update winit dependency to 0.29.
+
 ## [0.1.2](https://github.com/dbalsom/martypc/releases/tag/0.1.2) (2023-06-29)
 
 * Relicensed MartyPC under the MIT license.